Authentication and File upload functionality

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,19 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 
-class PostsController extends Controller
-{
+class PostsController extends Controller {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -35,14 +44,35 @@ class PostsController extends Controller
     public function store(Request $request) {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle file uppload
+        if ($request->hasFile('cover_image')) {
+            // Get the filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get only Filname
+            $filname = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get only Ext 
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            // Filename to store w/o any file name conflicts
+            $filenameToStore = $filname.'_'.time().'.'.$extension;
+
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/post_images',$filenameToStore);
+
+        } else{
+            $filenameToStore = 'noimage.jpg';
+        }
 
         // Creat post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $filenameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -67,6 +97,12 @@ class PostsController extends Controller
      */
     public function edit($id) {
         $post = Post::find($id);
+
+        //Check for the right user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect()->to('posts/'.$id)->with('error', '&#128683; Unauthorized access!&nbsp; You are not the owner, can\'t Edit the post');
+        }
+
         return view('posts.edit')->with('post',$post);
     }
 
@@ -80,13 +116,36 @@ class PostsController extends Controller
     public function update(Request $request, $id) {
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
+
         ]);
+        
+        // Handle file uppload
+        if ($request->hasFile('cover_image')) {
+            // Get the filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get only Filname
+            $filname = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get only Ext 
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+
+            // Filename to store w/o any file name conflicts
+            $filenameToStore = $filname.'_'.time().'.'.$extension;
+
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/post_images',$filenameToStore);
+
+        }
 
         // Update post
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        // Check for image
+        if ($request->hasFile('cover_image')) {
+            $post->cover_image = $filenameToStore;
+        }
         $post->save();
 
         return redirect()->to('posts/'.$id)->with(['success' => 'Post Updated']);
@@ -101,6 +160,17 @@ class PostsController extends Controller
     public function destroy($id) {
         // Delete post
         $post = Post::find($id);
+
+        //Check for the right user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect()->to('posts/'.$id)->with('error', '&#128683; Unauthorized access!&nbsp; You are not the owner, can\'t Delete the post');
+        }
+
+        //Delete the image (keep if noimage.jpg)
+        if ($post->cover_image != 'noimage.jpg') {
+            Storage::delete('public/post_images/'.$post->cover_image);
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('error', 'Post Removed');
